requirements.txt
transformers==3.0.2
dash==2.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
FastNLP==0.7.0
Flask==1.1.4
Flask-Compress==1.12
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SocketIO==5.2.0
eventlet==0.33.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
python-engineio==4.3.2
python-socketio==5.6.0
six==1.16.0
Werkzeug==1.0.1
pyrouge==0.1.3
rouge==1.0.1
